+++ utils.js
@@ -50,7 +50,41 @@ define(function (require, exports) {
         return promise;
     };
 
-    exports.show_popup_message = function (message) {
+    function get_merge_view_height(merge_view) {
+        function editorHeight(editor) {
+            if (!editor) {
+                return 0;
+            }
+            return editor.getScrollInfo().height;
+        }
+        return Math.max(editorHeight(merge_view.leftOriginal()),
+            editorHeight(merge_view.editor()),
+            editorHeight(merge_view.rightOriginal()));
+    }
+
+    exports.resize_merge_view = function (merge_view) {
+        var height = get_merge_view_height(merge_view);
+        for (;;) {
+            if (merge_view.leftOriginal()) {
+                merge_view.leftOriginal().setSize(null, height);
+            }
+            merge_view.editor().setSize(null, height);
+            if (merge_view.rightOriginal()) {
+                merge_view.rightOriginal().setSize(null, height);
+            }
+
+            var new_height = get_merge_view_height(merge_view);
+            if (new_height >= height) {
+                break;
+            }
+            else {
+                height = new_height;
+            }
+        }
+        merge_view.wrap.style.height = height + 'px';
+    };
+
+    exports.show_popup_message = function (message, disable_auto_hide) {
         var _$indicator = $('#brackets-nodejs-integration-runner-indicator');
         _$indicator.twipsy('hide').removeData('twipsy');
         var options = {
@@ -62,11 +96,13 @@ define(function (require, exports) {
             }
         };
         _$indicator.twipsy(options).twipsy('show');
-        setTimeout(function () {
-            if (_$indicator.data('twipsy')) {
-                _$indicator.twipsy('hide').removeData('twipsy');
-            }
-        }, 5000);
+        if (!disable_auto_hide) {
+            setTimeout(function () {
+                if (_$indicator.data('twipsy')) {
+                    _$indicator.twipsy('hide').removeData('twipsy');
+                }
+            }, 5000);
+        }
     };
 
     exports.uuid = function () {